move ProductEntityFieldMapper::getFlags() from project-base to the frontend-api package
